requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# AI/ML dependencies
openai>=1.3.0
langchain>=0.1.0

# Data processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Web scraping
lxml>=4.9.0
html5lib>=1.1

# Utilities
pyyaml>=6.0.1
colorama>=0.4.6

# Optional: for PDF generation
reportlab>=4.0.0

